refactor : asis 제거
