--- src/main/java/core/mvc/asis/RegacyHandlerMapping.java
@@ -1,33 +1,0 @@
-package core.mvc.asis;
-
-import core.mvc.HandlerMapping;
-import next.controller.asis.*;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.servlet.http.HttpServletRequest;
-import java.util.HashMap;
-import java.util.Map;
-
-public class RegacyHandlerMapping implements HandlerMapping {
-    private static final Logger logger = LoggerFactory.getLogger(DispatcherServlet.class);
-    private Map<String, Controller> mappings = new HashMap<>();
-
-    void initMapping() {
-
-        logger.info("Initialized Request Mapping!");
-        mappings.keySet().forEach(path -> {
-            logger.info("Path : {}, Controller : {}", path, mappings.get(path).getClass());
-        });
-    }
-
-    void put(String url, Controller controller) {
-        mappings.put(url, controller);
-    }
-
-    @Override
-    public Controller getHandler(HttpServletRequest request) {
-        final String url = request.getRequestURI();
-        return mappings.get(url);
-    }
-}
