--- src/main/java/core/mvc/tobe/MyController.java
@@ -1,35 +1,0 @@
-package core.mvc.tobe;
-
-import core.annotation.web.Controller;
-import core.annotation.web.RequestMapping;
-import core.annotation.web.RequestMethod;
-import core.mvc.JspView;
-import core.mvc.ModelAndView;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-@Controller
-public class MyController {
-    private static final Logger logger = LoggerFactory.getLogger(MyController.class);
-
-    @RequestMapping("/users")
-    public ModelAndView list(HttpServletRequest request, HttpServletResponse response) {
-        logger.debug("users findUserId");
-        return new ModelAndView(new JspView("/user/list.jsp"));
-    }
-
-    @RequestMapping(value="/users/show", method= RequestMethod.GET)
-    public ModelAndView show(HttpServletRequest request, HttpServletResponse response) {
-        logger.debug("users findUserId");
-        return new ModelAndView(new JspView("/user/profile.jsp"));
-    }
-
-    @RequestMapping(value="/users", method=RequestMethod.POST)
-    public ModelAndView create(HttpServletRequest request, HttpServletResponse response) {
-        logger.debug("users create");
-        return new ModelAndView(new JspView("redirect:/users"));
-    }
-}
