+++ src/main/java/next/controller/UserController.java
@@ -1,4 +1,4 @@
-package next.controller.tobe;
+package next.controller;
 
 import core.annotation.web.Controller;
 import core.annotation.web.RequestMapping;
@@ -6,7 +6,6 @@ import core.annotation.web.RequestMethod;
 import core.db.DataBase;
 import core.mvc.JspView;
 import core.mvc.ModelAndView;
-import next.controller.UserSessionUtils;
 import next.model.User;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
